requirements.txt
aiohttp==3.10.11
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.1
loguru==0.7.3
numpy==1.24.4
openai==1.104.0
packaging==25.0
pluggy==1.5.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.4
rsa==4.9.1
scikit-learn==1.3.2
sniffio==1.3.1
starlette==0.44.0
Telethon==1.41.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
